Support multi-gigabyte file sizes
Convert the type of the fileSize property from Int to BigInt (in the
DB) or Long (in the Scala code). This will let the API accept larger
file sizes.

+++ core/src/main/scala/uk/gov/nationalarchives/tdr/api/core/db/dao/FileDao.scala
@@ -54,7 +54,7 @@ class FileTable(tag: Tag) extends Table[FileRow](tag, "file") {
   def id = column[UUID]("id", O.PrimaryKey, O.AutoInc)
   def path = column[String]("path")
   def consignmentId = column[Int]("consignment_id")
-  def fileSize = column[Int]("file_size")
+  def fileSize = column[Long]("file_size")
   def lastModifiedDate = column[Instant]("last_modified_date")
   def fileName = column[String]("file_name")
   def consignment = foreignKey("file_consignment_fk", consignmentId, consignments)(_.id)
