+++ core/src/main/scala/uk/gov/nationalarchives/tdr/api/core/db/model/FileRow.scala
@@ -3,4 +3,4 @@ package uk.gov.nationalarchives.tdr.api.core.db.model
 import java.time.Instant
 import java.util.UUID
 
-case class FileRow (id: Option[UUID] = None, path: String, consignmentId: Int, fileSize: Int, lastModifiedDate: Instant, fileName: String)
+case class FileRow (id: Option[UUID] = None, path: String, consignmentId: Int, fileSize: Long, lastModifiedDate: Instant, fileName: String)
